+++ functions/scraping-events/src/scraping_events/main_cli.py
@@ -4,7 +4,7 @@ from argparse import ArgumentParser
 
 from scraping_events.logging_config import set_logging_config
 from scraping_events.playwright_utils import launch_browser
-from scraping_events.response import ResponseError, ResponseSuccess
+from scraping_events.schemas import ResponseError, ResponseSuccess
 from scraping_events.scrape_events import scrape_events
 
 set_logging_config()  # not main-guarded, so it's inherited by subprocesses
