+++ functions/scraping-events/src/scraping_events/schemas.py
